+++ main.go
@@ -15,9 +15,9 @@ import (
 	"github.com/pushrax/chihaya/config"
 	"github.com/pushrax/chihaya/server"
 
+	_ "github.com/pushrax/chihaya/storage/backend/batter"
+	_ "github.com/pushrax/chihaya/storage/backend/gazelle"
 	_ "github.com/pushrax/chihaya/storage/tracker/redis"
-	_ "github.com/pushrax/chihaya/storage/web/batter"
-	_ "github.com/pushrax/chihaya/storage/web/gazelle"
 )
 
 var (
