+++ server/stats_test.go
@@ -13,8 +13,8 @@ import (
 
 	"github.com/pushrax/chihaya/config"
 
+	_ "github.com/pushrax/chihaya/storage/backend/batter"
 	_ "github.com/pushrax/chihaya/storage/tracker/redis"
-	_ "github.com/pushrax/chihaya/storage/web/batter"
 )
 
 func newTestServer() (*Server, error) {
