+++ storage/backend/batter/batter.go
@@ -11,14 +11,14 @@ import (
 	"fmt"
 
 	"github.com/pushrax/chihaya/config"
-	"github.com/pushrax/chihaya/storage/web"
+	"github.com/pushrax/chihaya/storage/backend"
 
 	_ "github.com/bmizerany/pq"
 )
 
 type driver struct{}
 
-func (d *driver) New(conf *config.DataStore) web.Conn {
+func (d *driver) New(conf *config.DataStore) backend.Conn {
 	dsn := fmt.Sprintf(
 		"host=%s user=%s password=%s dbname=%s",
 		conf.Host,
@@ -47,10 +47,10 @@ func (c *Conn) Start() error {
 	return nil
 }
 
-func (c *Conn) RecordAnnounce(delta *web.AnnounceDelta) error {
+func (c *Conn) RecordAnnounce(delta *backend.AnnounceDelta) error {
 	return nil
 }
 
 func init() {
-	web.Register("batter", &driver{})
+	backend.Register("batter", &driver{})
 }
